edit

+++ app/Http/Controllers/Api/CategoryController.php
@@ -61,6 +61,9 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::find($id);
+        if ($category->books()->count() > 0) 
+             return ResponseHelper::failed("لا يمكن حذف الصنف لوجود كتب مرتبطة به" , null);
+    
         $category->delete();
         return ResponseHelper::success("تم حذف الصنف" , $category);
     }
